Peers are now selectable and app doesn't crash on nearby train modal pop up

+++ Songtrain/ViewController.h
@@ -43,7 +43,7 @@
 @property UICollectionView *nearbyTrainsModal;
 @property UIView *nearbyTrainBackground;
 
-@property int lastIndex;
+@property NSInteger lastIndex;
 
 -(void)updatePlayOrPauseImage;
